refactor(form-builder): hoist useOnClickOutside hook to shared directory + add comment about why its needed
