+++ packages/@sanity/form-builder/src/hooks/useOnClickOutside.ts
@@ -1,6 +1,9 @@
-// Modified version of https://usehooks.com/useOnClickOutside that can take multiple refs
 import {RefObject, useEffect} from 'react'
 
+// This is a modified version of https://usehooks.com/useOnClickOutside that can take multiple element refs
+// There is currently a bug in the `useClickOutside` hook from @sanity/ui that requires the refs to be passed as
+// actual HTML elements instead of mutable refs. This requires the consumer to store elements in component state
+// which adds quite a bit of cruft and isn't always feasible
 export function useOnClickOutside(refs: RefObject<HTMLElement>[], handler) {
   useEffect(() => {
     const listener = (event) => {
