+++ packages/@sanity/form-builder/src/inputs/ReferenceInput/ArrayItemReferenceInput.tsx
@@ -61,6 +61,7 @@ import {DragHandle} from '../arrays/common/DragHandle'
 import {InsertEvent} from '../arrays/ArrayOfObjectsInput/types'
 import randomKey from '../arrays/common/randomKey'
 import {InsertMenu} from '../arrays/ArrayOfObjectsInput/InsertMenu'
+import {useOnClickOutside} from '../../hooks/useOnClickOutside'
 import {BaseInputProps, CreateOption, SearchState} from './types'
 import {OptionPreview} from './OptionPreview'
 import {useReferenceInfo} from './useReferenceInfo'
@@ -68,7 +69,6 @@ import {PreviewReferenceValue} from './PreviewReferenceValue'
 import {CreateButton} from './CreateButton'
 import {ReferenceAutocomplete} from './ReferenceAutocomplete'
 import {AutocompleteContainer} from './AutocompleteContainer'
-import {useOnClickOutside} from './utils/useOnClickOutside'
 
 const StyledPreviewCard = styled(PreviewCard)`
   /* this is a hack to avoid layout jumps while previews are loading
